+++ lib/src/effects/slide_effect.dart
@@ -6,6 +6,11 @@ import 'indicator_effect.dart';
 
 /// Holds painting configuration to be used by [SlidePainter]
 class SlideEffect extends BasicIndicatorEffect {
+  /// The effect variant
+  ///
+  /// defaults to [SlideType.normal]
+  final SlideType type;
+
   /// Default constructor
   const SlideEffect({
     Color activeDotColor = Colors.indigo,
@@ -17,6 +22,7 @@ class SlideEffect extends BasicIndicatorEffect {
     Color dotColor = Colors.grey,
     double strokeWidth = 1.0,
     PaintingStyle paintStyle = PaintingStyle.fill,
+    this.type = SlideType.normal,
   }) : super(
             dotWidth: dotWidth,
             dotHeight: dotHeight,
@@ -31,4 +37,12 @@ class SlideEffect extends BasicIndicatorEffect {
   IndicatorPainter buildPainter(int count, double offset) {
     return SlidePainter(count: count, offset: offset, effect: this);
   }
+}
+
+/// The Slide effect variants
+enum SlideType {
+  /// Paints regular dot sliding animation
+  normal,
+  /// Paints masked (under-layered) dot sliding animation
+  slideUnder
 }