+++ lib/src/painters/worm_painter.dart
@@ -24,7 +24,8 @@ class WormPainter extends BasicIndicatorPainter {
     // paint still dots
     paintStillDots(canvas, size);
 
-    final activeDotPaint = Paint()..color = effect.activeDotColor;
+    final activeDotPaint = Paint()
+      ..color = effect.activeDotColor;
     final dotOffset = (offset - offset.toInt());
 
     // handle dot travel from end to start (for infinite pager support)
@@ -47,7 +48,7 @@ class WormPainter extends BasicIndicatorPainter {
     var head = xPos;
     var tail = xPos + effect.dotWidth + (wormOffset * distance);
     var halfHeight = effect.dotHeight / 2;
-    var thinWorm = effect.type == WormType.thin;
+    final thinWorm = effect.type == WormType.thin || effect.type == WormType.thinUnderground;
     var dotHeight = thinWorm ? halfHeight + (halfHeight * (1 - wormOffset)) : effect.dotHeight;
 
     if (wormOffset > 1) {
@@ -64,6 +65,13 @@ class WormPainter extends BasicIndicatorPainter {
       yPos + dotHeight / 2,
       dotRadius,
     );
-    canvas.drawRRect(worm, activeDotPaint);
+    if (effect.type == WormType.underground || effect.type == WormType.thinUnderground) {
+      canvas.saveLayer(Rect.largest, Paint());
+      canvas.drawRRect(worm, activeDotPaint);
+      maskStillDots(size, canvas);
+      canvas.restore();
+    } else {
+      canvas.drawRRect(worm, activeDotPaint);
+    }
   }
 }